PAAC-154: Renamed calculators.period package to calculators.internal
