+++ app/calculators/internal/Period2Calculator.scala
@@ -14,11 +14,11 @@
  * limitations under the License.
  */
 
-package calculators.periods
+package calculators.internal
 
 import models._
 import calculators.results.SummaryResultCalculator
-import calculators.periods.Utilities._
+import calculators.internal.Utilities._
 import calculators.Utilities._
 
 protected class Period2Calculator(implicit allowanceInPounds: Long,
