+++ app/calculators/results/Year2015Period2Calculator.scala
@@ -17,8 +17,7 @@
 package calculators.results
 
 import models._
-import calculators.periods._ 
-import calculators.AllowanceCalculator
+import calculators.internal._
 
 /**
   Calculator for 2015, period 2 (from 9th July 2015 to April 2016)
